Adapt manifests for the development environment
Also tag the development image as the production image

This will be used occasionally when we need to refer to sles12/velum:0.0
as happens in production.

Add dev-env manifest that will take care of creating databases for
development and test environments, now that we aren't connecting with
the `root` mysql user and our permissions are restricted.